Deduplicate XCM decoding logic
