+++ cumulus/pallets/xcmp-queue/src/benchmarking.rs
@@ -22,7 +22,6 @@ use codec::DecodeAll;
 use frame_benchmarking::v2::*;
 use frame_support::traits::Hooks;
 use frame_system::RawOrigin;
-use xcm::v3::MAX_INSTRUCTIONS_TO_DECODE;
 
 #[benchmarks]
 mod benchmarks {
