+++ polkadot/xcm/src/lib.rs
@@ -46,11 +46,17 @@ pub mod latest {
 mod double_encoded;
 pub use double_encoded::DoubleEncoded;
 
+mod utils;
+
 #[cfg(test)]
 mod tests;
 
 /// Maximum nesting level for XCM decoding.
 pub const MAX_XCM_DECODE_DEPTH: u32 = 8;
+/// The maximal number of instructions in an XCM before decoding fails.
+///
+/// This is a deliberate limit - not a technical one.
+pub const MAX_INSTRUCTIONS_TO_DECODE: u8 = 100;
 
 /// A version of XCM.
 pub type Version = u32;
@@ -614,9 +620,12 @@ fn size_limits() {
 
 #[test]
 fn validate_xcm_nesting_works() {
-	use crate::latest::{
-		prelude::{GeneralIndex, ReserveAssetDeposited, SetAppendix},
-		Assets, Xcm, MAX_INSTRUCTIONS_TO_DECODE, MAX_ITEMS_IN_ASSETS,
+	use crate::{
+		latest::{
+			prelude::{GeneralIndex, ReserveAssetDeposited, SetAppendix},
+			Assets, Xcm, MAX_ITEMS_IN_ASSETS,
+		},
+		MAX_INSTRUCTIONS_TO_DECODE,
 	};
 
 	// closure generates assets of `count`
